requirements.txt
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
bcrypt==4.3.0
black==24.8.0
cffi==1.17.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
ConfigSpace==1.2.1
contourpy==1.1.1
cryptography==45.0.5
cycler==0.12.1
dask==2023.5.0
dask-jobqueue==0.8.5
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
distributed==2023.5.0
emcee==3.1.6
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.57.0
fsspec==2025.3.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
locket==1.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
more-itertools==10.5.0
msgpack==1.1.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==1.24.4
packaging==25.0
pandas==2.0.3
paramiko==3.2.0
parso==0.8.4
partd==1.4.1
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
POAP==0.1.26
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyDOE==0.3.8
pyDOE2==1.3.0
Pygments==2.19.1
PyNaCl==1.5.0
pynisher==0.6.4
pyparsing==3.1.4
pyrfr==0.9.0
pySOT==0.3.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
smac==1.4.0
sortedcontainers==2.4.0
stack-data==0.6.3
tblib==3.0.0
threadpoolctl==3.5.0
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
utils==1.0.2
wcwidth==0.2.13
zict==3.0.0
zipp==3.20.2
